requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
google-cloud-storage==2.19.0
gunicorn==21.2.0
pytest==7.4.0
google-generativeai==0.8.4
google-ai-generativelanguage==0.6.15
google-api-python-client==2.166.0
